[SDPA-2332] Move ImageGallery into no-ssr folder.
