+++ packages/Organisms/ImageGallery/no-ssr/index.js
@@ -2,8 +2,14 @@
 // Provide a plugin which others can use it in Nuxt as no ssr mode.
 // Usage in Nuxt: https://nuxtjs.org/guide/plugins/#client-side-only
 
-import RplImageGalleryModal from './ImageGalleryModal.vue'
-import RplFullscreenImage from './FullscreenImage.vue'
+import RplImageGallery from './ImageGallery.vue'
 
-export { RplImageGalleryModal }
-export { RplFullscreenImage }
+const install = Vue => {
+  Vue.component('rplImageGallery', RplImageGallery)
+}
+
+export default {
+  install
+}
+
+export { RplImageGallery }
