+++ src/storybook-components/PrintPage.vue
@@ -132,7 +132,7 @@ import RplCampaignSecondary from '@dpc-sdp/ripple-campaign-secondary'
 import { RplLatestEvents } from '@dpc-sdp/ripple-event'
 import { RplGrantsOverview, RplGrantsListItem } from '@dpc-sdp/ripple-grants'
 import { RplHeroBanner, RplIntroBanner, RplHeroBannerCta } from '@dpc-sdp/ripple-hero-banner'
-import { RplImageGallery } from '@dpc-sdp/ripple-image-gallery'
+import { RplImageGallery } from '@dpc-sdp/ripple-image-gallery/no-ssr'
 import { RplMarkup } from '@dpc-sdp/ripple-markup'
 import { RplFeaturedNews, RplNewsListing } from '@dpc-sdp/ripple-news'
 import { RplProfileHighlight, RplProfileHighlightHonourRoll } from '@dpc-sdp/ripple-profile'
